posix: move some utility functions to new file
