+++ src/libpmemfile-posix/callbacks.c
@@ -35,11 +35,10 @@
  */
 
 #include "callbacks.h"
+#include "compiler_utils.h"
 #include "internal.h"
 #include "os_thread.h"
 #include "out.h"
-#include "util.h"
-
 #include <errno.h>
 
 struct tx_callback {
